Release: pluggable AI provider, server-side catalogue, bounded pagination

+++ schemas/transaction/input_dto.py
@@ -1,13 +1,16 @@
+from decimal import Decimal
+
 from pydantic import BaseModel, Field
+
 from core.enum import TransactionType
 
 
 class TransactionCreateDTO(BaseModel):
-    amount: float = Field(..., gt=0)
+    amount: Decimal = Field(..., gt=0, max_digits=10, decimal_places=2)
     type: TransactionType
     description: str
     client_fk: int
 
 
 class PaymentRequestDTO(BaseModel):
-    amount: float = Field(..., gt=0)
+    amount: Decimal = Field(..., gt=0, max_digits=10, decimal_places=2)
